Refactor lazy loading options and error handling
