+++ src/defaultOptions.js
